# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAPhoto
  description: >-
    Keep in touch with your friends
    by sharing photos of special moments,
    thanks to WASAPhoto!

    You can upload your photos directly from your PC,
    and they will be visible to everyone following you.
  version: 0.0.1
...
components:
  schemas:
    UserID:
      description: UserID.
      readOnly: true
      type: number
      format: uint64
      example: 1
    PhotoID:
      description: PhotoID.
      readOnly: true
      type: number
      format: uint64
      example: 1
    CommentID:
      description: CommentID.
      readOnly: true
      type: number
      format: uint64
      example: 1
    Username:
      description: Username.
      type: string
      pattern: '^.*?$'
      minLength: 3
      maxLength: 16
      example: Maria
    User:
      description: User object.
      type: object
      properties:
        userID:
          $ref: '#/components/schemas/UserID'
        username:
          $ref: '#/components/schemas/Username'
    Date:
      description: Publish date.
      type: string
      format: date-time
      minLength: 19
      maxLength: 19
      example: '2006-01-02 15:04:05'
    Photo:
      description: Photo object.
      type: object
      properties:
        photoID:
          $ref: '#/components/schemas/PhotoID'
        authorID:
          $ref: '#/components/schemas/UserID'
        format:
          description: Photo format.
          type: string
          enum:
            - png
            - jpg
          pattern: '^.*?$'
          minLength: 3
          maxLength: 3
          example: png
        date:
          $ref: '#/components/schemas/Date'
        likesList:
          description: Users list.
          type: array
          minItems: 0
          maxItems: 99
          items: { $ref: '#/components/schemas/User' }
        commentsList:
          description: Comments list.
          type: array
          minItems: 0
          maxItems: 99
          items: { $ref: '#/components/schemas/Comment' }
    CommentText:
          description: Comment text.
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 2200
          example: Amazing photo!
    Comment:
      description: Comment object.
      type: object
      properties:
        commentID:
          $ref: '#/components/schemas/CommentID'
        photoID:
          $ref: '#/components/schemas/PhotoID'
        authorID:
          $ref: '#/components/schemas/UserID'
        commentText:
          $ref: '#/components/schemas/CommentText'
        date:
          $ref: '#/components/schemas/Date'
...
  parameters:
    UserID:
      name: uid
      in: path
      required: true
      description: UserID.
      schema:
        $ref: '#/components/schemas/UserID'
    PhotoID:
      name: pid
      in: path
      required: true
      description: PhotoID.
      schema:
        $ref: '#/components/schemas/PhotoID'
...
    Username:
      name: username
      in: path
      required: true
      description: Username.
      schema:
        $ref: '#/components/schemas/Username'
  responses:
    Unauthorized:
      # 401
      description: The client must authenticate.
    Forbidden:
      # 403
      description: Some user has banned the other.
    InternalServerError:
      #500
      description: Server error.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []
paths:
...
  /settings:

    put:
      operationId: setMyUserName
      tags: [settings]
      summary: update username
      description: Update username.
      requestBody:
        description: Updated username.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/User' }
      responses:
        "200":
          # OK
          description: |-
            Username successfully updated.
            The updated user is returned in the content.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        "400":
          # Bad Request
          description: |-
            Invalid JSON object or;
            invalid username.
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'
  /following/{uid}:

    parameters:
      - $ref: '#/components/parameters/UserID'

    put:
      operationId: followUser
      tags: [following]
      summary: start following a user
      description: Start following a user.
      responses:
        "204":
          # No Content
          description: User started followin other user.
        "400":
          # Bad Request
          description: |-
            Invalid path parameter uid or;
            uid not matching any existing user or;
            user trying to follow himself or;
            user already following.
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "403":
          # Forbidden
          $ref: '#/components/responses/Forbidden'
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'

    delete:
      operationId: unfollowUser
      tags: [following]
      summary: unfollow a user
      description: Unfollow a user.
      responses:
        "204":
          # No Content
          description: User unfollowed other user.
        "400":
          # Bad Request
          description: |-
            Invalid path parameter uid or;
            uid not matching any existing user.
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'
  /banned/{uid}:

    parameters:
      - $ref: '#/components/parameters/UserID'

    put:
      operationId: banUser
      tags: [banned]
      summary: ban a user
      description: Ban a user.
      responses:
        "204":
          # No Content
          description: User banned other user.
        "400":
          # Bad Request
          description: |-
            Invalid path parameter uid or;
            uid not matching any existing user or;
            user trying to ban himself or;
            user already banned.
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'

    delete:
      operationId: unbanUser
      tags: [banned]
      summary: unban a user
      description: Unban a user.
      responses:
        "204":
          # No Content
          description: User unbanned other user.
        "400":
          # Bad Request
          description: |-
            Invalid path parameter uid or;
            uid not matching any existing user.
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'
...
  /photos/{pid}/comments/:

    parameters:
      - $ref: '#/components/parameters/PhotoID'

    post:
      operationId: commentPhoto
      tags: [comments]
      summary: comment on a photo
      description: Comment on a photo.
      requestBody:
        description: Comment text.
        required: true
        content:
          text/plain:
            schema: { $ref: '#/components/schemas/CommentText' }
      responses:
        "201":
          # Created
          description: Comment successfully added to the specified photo.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Comment' }
        "400":
          # Bad Request
          description: |-
            Invalid path parameter pid or;
            pid not matching any existing photo or;
            invalid comment data.
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "403":
          # Forbidden
          $ref: '#/components/responses/Forbidden'
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'

    get:
      operationId: getCommentsList
      tags: [getters]
      summary: get photo comments
      description: Get the list of comments for a specific photo.
      responses:
        "200":
          # OK
          description: List of photo comments retrieved successfully.
          content:
            application/json:
              schema:
                description: Comments list.
                type: array
                minItems: 0
                maxItems: 99
                items: { $ref: '#/components/schemas/Comment' }
        "400":
          # Bad Request
          description: |-
            Invalid path parameter pid or;
            pid not matching any existing photo.
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "403":
          # Forbidden
          $ref: '#/components/responses/Forbidden'
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'
...
  /photos/{pid}/likes/:

    parameters:
      - $ref: '#/components/parameters/PhotoID'

    get:
      operationId: getLikesList
      tags: [getters]
      summary: get photo likes
      description: Get the list of users who liked a specific photo.
      responses:
        "200":
          # OK
          description: List of photo likes retrieved successfully.
          content:
            application/json:
              schema:
                description: Users list.
                type: array
                minItems: 0
                maxItems: 99
                items: { $ref: '#/components/schemas/User' }
        "400":
          # Bad Request
          description: |-
            Invalid path parameter pid or;
            pid not matching any existing photo.
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "403":
          # Forbidden
          $ref: '#/components/responses/Forbidden'
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'
  /stream:

    get:
      operationId: getMyStream
      tags: [getters]
      summary: get user stream
      description: Get the stream of photos and activities for the authenticated user.
      responses:
        "200":
          # OK
          description: User stream retrieved successfully.
          content:
            application/json:
              schema:
                description: Photos list.
                type: array
                minItems: 0
                maxItems: 99
                items: { $ref: '#/components/schemas/Photo' }
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'
  /user/{username}:

    parameters:
      - $ref: '#/components/parameters/Username'

    get:
      operationId: getUserId
      tags: [getters]
      summary: get user ID
      description: Get user ID by specifying username.
      responses:
        "200":
          # OK
          description: User ID retrieved successfully.
          content:
            text/plain:
              schema: { $ref: '#/components/schemas/UserID' }
        "400":
          # Bad Request
          description: |-
            Invalid path parameter username or;
            username not matching any existing user.
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'
  /users/:

    get:
      operationId: searchUserByUsername
      tags: [search]
      summary: search a user
      description: >-
        Search a user by providing a username (might be partial).
        A list of users partially matching the provided username is returned.
      parameters:
        - name: username
          in: query
          required: true
          description: Username (might be partial).
          schema: { $ref: '#/components/schemas/Username' }
      responses:
        "200":
          # OK
          description: List of matching users retrieved successfully.
          content:
            application/json:
              schema:
                description: Users list.
                type: array
                minItems: 0
                maxItems: 99
                items: { $ref: '#/components/schemas/User' }
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'